Added androidNavigationBar configuration so the default navigation buttons and status icons are hidden
Added a styled back button to the information and Settings screen but commented it out for now. Clicking on a tab icon is considered the root of navigation. The back button needs to be there on screens a layer deeper. On the root tab screen, it will always navigate back to Milestones since this is the initial screen.

Changed background color in dark mode so it's not the same as the header