Made the MilestoneScreen zoomable using react-native-zoomable-view
Added touch handling that logs where on the screen the user is pressing.

Added flag recognition using the actual x, y, width, height values of the flags on the MilestoneScreen. Data taken from Figma. This makes it so that each flag has a defined area. This way the app knows when the user pressed on a specific flag. No actions besides a log have been added yet

Added a 2000px SVG and png version of the Milestone-Background (might still get removed later)