Merge user-stories 1 & 6 into main (#21)
* Changed the splash screen icon

Added placeholder screens for Milestones, Information and Settings

Added basic tab navigation

* Added icons for the tab navigation screens.

Added dynamic coloring & themes, so the app can easily switch between light and dark mode styling (colors in theme.ts will still be adjusted)

Added react native svg transformer support.

* Added androidNavigationBar configuration so the default navigation buttons and status icons are hidden

Added a styled back button to the information and Settings screen but commented it out for now. Clicking on a tab icon is considered the root of navigation. The back button needs to be there on screens a layer deeper. On the root tab screen, it will always navigate back to Milestones since this is the initial screen.

Changed background color in dark mode so it's not the same as the header

* Reversed the tab colors in light mode. The background is now white; the active tab is green and the inactive tabs are black

* dark mode theme color adjustment

* Made the MilestoneScreen zoomable using react-native-zoomable-view

Added touch handling that logs where on the screen the user is pressing.

Added flag recognition using the actual x, y, width, height values of the flags on the MilestoneScreen. Data taken from Figma. This makes it so that each flag has a defined area. This way the app knows when the user pressed on a specific flag. No actions besides a log have been added yet

Added a 2000px SVG and png version of the Milestone-Background (might still get removed later)