2 micro goals (#22)
* Nested a stack navigator for the milestones in the tab navigator. Now when pressing on flag 1, it takes you to a different screen. This screen will become an overview screen.

* Made ActiveColors an importable componenent.

Changed the tab bar, so it has a green background but the active tab has a white background.

Added Noto Serif as a usable font.

Started work on the Microgoals screen. Added custom svg's and a placeholder list view. This screen will be turned into a reusable component later.

* Implemented Typescript type checking for navigation. Since I've been using the Dynamic API to create my navigation, it is necessary to specify the types for each screen as well as the nesting structure as it cannot be inferred from the code (https://reactnavigation.org/docs/typescript/?config=dynamic#combining-navigation-props).

Added image files for the item containers

Implemented the new typechecked navigation in the screens from the MilestonesStack.

Fixed the go back button in the MilestonesStack

* Added some styling for the back button

Adjusted the colours from the themes

Build a static activity page. The next step is to make this a reusable component

* Made the ActivityScreen a reusable component that accepts props and adjusts the concent based on the parameters it receives

* Added an aditional contentExtra prop for an activitypage, that displays text that needs to appear under the image.

Added a switch case for styling the activity pages based on the given category

Added a background for the title container

Made some reusable buttons (primary and secondary)