3 milestone progress tracking (#23)
* Refactored the code to make the MicrogoalsOverviewScreen a reusable screen that accepts params and renders items based on received data.

The flags on the main MilestonesScreen have been linked with a milestone. When pressed, the flag navigates to the MicrogoalOverviewScreen and passes along the corresponding milestone id. The data then gets fetched based on the id and used to fill the content.

* Added async storage for storing progress in the background.

Made an entire item clickable, not just the title button.

Made it possible to track which activities have been completed and which haven't. Using Context and a provider, we can store completed activities in the local storage. We get all stored items when opening a milestone (based on id), and mark activities that have been stored already as complete.

The user gets send back to the main milestones screen after completing an activity. I would like to add a transition screen between these two.

The context is wrapped around the root app layout

* First steps of filling out the content. I'm moving this to a new branch