Changed the activity progress tracking structure so it takes the milestone id as well.
Added a static total number of activities to the milestoneFlags.

Added a way to count the number of completed activities from a certain milestone. The percentage of this (completed activities / total activities) gets displayed on the main screen.

Added React Reanimted config.

Added a completion screen that briefly shows after completing an activity.

Added an activity indicator while the milestonesScreen is being rendered.