4 5 language and integration micro goals (#27)
* Added milestone 3 data

Added content for all first 3 milestones, including a new work topic for the activities

Added accessiblity labels

* Removed the static totalActivities variable from the flag mapping. The number of activities in the data object now get counted with a count function.

Added a button on milestone screen that looks for the nearest uncompleted milestone and navigates to it. This helps with UX.

Added a temporary button on the settings screen that wipes the entire async storage. This is for user testing.

* Added more images to the activities